trying to implement scroll with existing utility scrollToID
