+++ src/utilities.js
@@ -200,6 +200,8 @@ export const scrollToTop = (config = {}) => {
 };
 
 export const scrollToId = ({ el, config = {} }) => {
+  console.log('in scrollToId');
+  console.log('el: ', el);
   // Get an element's distance from the top of the page
   const getElemDistance = (elem) => {
     let location = 0;
@@ -212,6 +214,7 @@ export const scrollToId = ({ el, config = {} }) => {
     return location >= 0 ? location : 0;
   };
   const elem = document.querySelector(el);
+  console.log('elem: ', elem);
   const location = getElemDistance(elem);
 
   scrollTo(location, config);
